fix(bg-jobs): Fix running once when no job was scheduled
